Add a requirements-dev.txt file for CI related tools
This merge adds the list of tools used to lint, format, and check
the repository code. Note that this does not include the actual
dependencies required to execute the code in the repository at
runtime.

TODO: Add a similar requirements.txt file to enable dependency
installation by a user wishing to install and use the SPIIR package.